Moving replay server and pixel tester under test directory
