+++ src/test/server/controllers/webhook-events/status-test.ts
@@ -3,7 +3,6 @@ import {Server} from 'http';
 import * as sinon from 'sinon';
 import {SinonSandbox, SinonStub} from 'sinon';
 
-import {startTestReplayServer} from '../../../../replay-server';
 import * as notificationController from '../../../../server/controllers/notifications';
 import {handleStatus} from '../../../../server/controllers/webhook-events/status';
 import {StatusHook} from '../../../../server/controllers/webhook-events/types';
@@ -15,6 +14,7 @@ import * as performAutomergeModule from '../../../../server/utils/perform-autome
 import {initFirestore} from '../../../../utils/firestore';
 import {initGithub} from '../../../../utils/github';
 import {initSecrets} from '../../../../utils/secrets';
+import {startTestReplayServer} from '../../../utils/replay-server';
 
 const TEST_SECRETS = {
   GITHUB_CLIENT_ID: 'ClientID',
