+++ src/test/server/util/my-repos-test.ts
@@ -1,8 +1,8 @@
 import test from 'ava';
 
-import {startTestReplayServer} from '../../../replay-server';
 import {getMyRepos} from '../../../server/utils/my-repos';
 import {initGithub} from '../../../utils/github';
+import {startTestReplayServer} from '../../utils/replay-server';
 
 test('top contributed repos for samuelli', async (t) => {
   const {server, url} = await startTestReplayServer(t);
